+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/pom.xml
@@ -166,5 +166,10 @@
       <artifactId>hamcrest-library</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.opennms.core.test-api</groupId>
+      <artifactId>org.opennms.core.test-api.xml</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
