+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/src/main/java/org/opennms/features/topology/plugins/topo/graphml/GraphmlRepositoryImpl.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.graphml.service;
+package org.opennms.features.topology.plugins.topo.graphml;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -40,6 +40,7 @@ import java.util.Properties;
 import javax.xml.bind.JAXB;
 
 import org.graphdrawing.graphml.GraphmlType;
+import org.opennms.features.graphml.service.GraphmlRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
