+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.graphml/src/test/java/org/opennms/features/topology/plugins/topo/graphml/GraphmlRepositoryImplTest.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.graphml.service;
+package org.opennms.features.topology.plugins.topo.graphml;
 
 import java.io.ByteArrayOutputStream;
 import java.io.FileInputStream;
