+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/v1/GraphMLRestService.java
@@ -30,6 +30,7 @@ package org.opennms.web.rest.v1;
 
 import java.io.IOException;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -39,24 +40,29 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
 
 import org.graphdrawing.graphml.GraphmlType;
+import org.opennms.core.soa.ServiceRegistry;
+import org.opennms.core.spring.BeanUtils;
 import org.opennms.features.graphml.model.GraphML;
 import org.opennms.features.graphml.model.GraphMLReader;
 import org.opennms.features.graphml.model.InvalidGraphException;
 import org.opennms.features.graphml.service.GraphmlRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
 @Path("graphml")
 public class GraphMLRestService {
 
-    @Autowired
-    private GraphmlRepository graphmlRepository;
+    private ServiceRegistry serviceRegistry;
 
     @POST
     @Path("{graph-name}")
     public Response createGraph(@PathParam("graph-name") String graphname,
                                 GraphmlType graphmlType) throws IOException {
+        // Get Service
+        final GraphmlRepository graphmlRepository = getServiceRegistry().findProvider(GraphmlRepository.class);
+        if (graphmlRepository == null) {
+            return temporarilyNotAvailable();
+        }
 
         // Verify that it does not already exist
         if (graphmlRepository.exists(graphname)) {
@@ -77,6 +83,12 @@ public class GraphMLRestService {
     @DELETE
     @Path("{graph-name}")
     public Response deleteGraph(@PathParam("graph-name") String graphname) throws IOException {
+        // Get Service
+        final GraphmlRepository graphmlRepository = getServiceRegistry().findProvider(GraphmlRepository.class);
+        if (graphmlRepository == null) {
+            return temporarilyNotAvailable();
+        }
+
         if (!graphmlRepository.exists(graphname)) {
             throw new NoSuchElementException("No GraphML file found with name  " + graphname);
         }
@@ -87,8 +99,29 @@ public class GraphMLRestService {
     @GET
     @Path("{graph-name}")
     public Response getGraph(@PathParam("graph-name") String graphname) throws IOException {
+        // Get Service
+        final GraphmlRepository graphmlRepository = getServiceRegistry().findProvider(GraphmlRepository.class);
+        if (graphmlRepository == null) {
+            return temporarilyNotAvailable();
+        }
+
         GraphmlType byName = graphmlRepository.findByName(graphname);
         return Response.ok(byName).build();
+    }
+
+    private ServiceRegistry getServiceRegistry() {
+        if (serviceRegistry == null) {
+            serviceRegistry = BeanUtils.getBean("soaContext", "serviceRegistry", ServiceRegistry.class);
+            Objects.requireNonNull(serviceRegistry);
+        }
+        return serviceRegistry;
+    }
+
+    private static Response temporarilyNotAvailable() {
+        return Response
+                .status(Response.Status.SERVICE_UNAVAILABLE)
+                .entity("No service registered to handle your query. This is a temporary issue. Please try again later.")
+                .build();
     }
 
 }
